support range  validation for decimal

+++ datavalidation.go
@@ -120,9 +120,9 @@ func (dd *DataValidation) SetDropList(keys []string) error {
 }
 
 // SetRange provides function to set data validation range in drop list.
-func (dd *DataValidation) SetRange(f1, f2 int, t DataValidationType, o DataValidationOperator) error {
-	formula1 := fmt.Sprintf("%d", f1)
-	formula2 := fmt.Sprintf("%d", f2)
+func (dd *DataValidation) SetRange(f1, f2 float64, t DataValidationType, o DataValidationOperator) error {
+	formula1 := fmt.Sprintf("%f", f1)
+	formula2 := fmt.Sprintf("%f", f2)
 	if dataValidationFormulaStrLen+21 < len(dd.Formula1) || dataValidationFormulaStrLen+21 < len(dd.Formula2) {
 		return fmt.Errorf(dataValidationFormulaStrLenErr)
 	}
